+++ app/models/concerns/processable/facebook_processor.rb
@@ -1,0 +1,10 @@
+module Processable
+  module FacebookProcessor
+    extend ActiveSupport::Concern
+    include Processable
+
+    included do
+
+    end
+  end
+end
